+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change history for ui-agreements
 
 ## 12.2.0 IN PROGRESS
+  * ERM-4014: Change default `headingLevel` in meta section header to `3`
 
 ## 12.1.0 2026-04-17
   * ERM-3975: Agreements|Local KB Package view: List item li does not have a ul parent element